Fix: Move data provider
