+++ test/Unit/Component/News/PublicationTest.php
@@ -33,7 +33,7 @@ class PublicationTest extends \PHPUnit_Framework_TestCase
     }
 
     /**
-     * @dataProvider Refinery29\Sitemap\Test\Unit\Component\DataProvider::providerInvalidString
+     * @dataProvider Refinery29\Sitemap\Test\Util\DataProvider\InvalidString::data
      *
      * @param mixed $name
      */
@@ -50,7 +50,7 @@ class PublicationTest extends \PHPUnit_Framework_TestCase
     }
 
     /**
-     * @dataProvider Refinery29\Sitemap\Test\Unit\Component\DataProvider::providerInvalidString
+     * @dataProvider Refinery29\Sitemap\Test\Util\DataProvider\InvalidString::data
      *
      * @param mixed $language
      */
